requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.43.3
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Eel==0.16.0
et-xmlfile==1.1.0
fastjsonschema==2.19.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.7
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pure-eval==0.2.2
pycparser==2.22
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
requests==2.31.0
setuptools==69.5.1
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
whichcraft==0.6.1
widgetsnbextension==4.0.10
zope.event==5.0
zope.interface==6.3

PyQt6~=6.7.0